Rework listenForPackageChanges to ensure unregistration
